Rename `buffers` module to `buffer`
... and move `StaticBuffer` to nested `static` module
