+++ wgpu/src/buffer.rs
@@ -1,0 +1,3 @@
+//! Utilities for buffer operations.
+pub mod dynamic;
+pub mod r#static;
